Ostalo da se poradi na View-u

+++ src/BoilerplateCore.Application/Services/ROsoba.cs
@@ -34,7 +34,11 @@ namespace BoilerplateCore.Services
 
         public OsobaDTO GetById(int id)
         {
-            Osoba osobeId = _osobaRepository.Get(id);
+            Osoba osobeId = _osobaRepository.FirstOrDefault(c => c.Id == id);
+            if (osobeId == null)
+            {
+                throw new UserFriendlyException("Osoba sa naznacenim Id ne posoji.");
+            }
             OsobaDTO osobaDto = _objectMapper.Map<OsobaDTO>(osobeId);
 
             return osobaDto;
@@ -42,19 +46,32 @@ namespace BoilerplateCore.Services
 
         public void Create(OsobaCreateDto osoba)
         {
+            var provOsobe = _osobaRepository.FirstOrDefault(c => c.Id == osoba.Id);
+            if (provOsobe != null)
+            {
+                throw new UserFriendlyException("Osoba sa naznacenim Id vec posoji.");
+            }
             var osobaC = _objectMapper.Map<Osoba>(osoba);
             _osobaRepository.Insert(osobaC);
         }
 
-        public void Update(OsobaCreateDto osoba)
+        public void Update(int id, OsobaCreateDto osoba)
         {
-            var osobaId = _osobaRepository.Get(osoba.Id);
+            var osobaId = _osobaRepository.FirstOrDefault(c => c.Id == osoba.Id);
+            if (osobaId == null)
+            {
+                throw new UserFriendlyException("Osoba sa naznacenim Id ne posoji.");
+            }
             _objectMapper.Map(osoba, osobaId);
         }
 
         public void Delete(int id)
         {
-            var osobaId = _osobaRepository.Get(id);
+            var osobaId = _osobaRepository.FirstOrDefault(c => c.Id == id);
+            if (osobaId==null)
+            {
+                throw new UserFriendlyException("Unijeli ste nepostojeci Id.");
+            }
             _osobaRepository.Delete(osobaId);
 
         }
